Updated enums back to int. added __packed attrubte so the sensoryPacket does not take up 88 bytes instead of 84bytes.

+++ src/Sannel/src/SensorTypes.h
@@ -21,7 +21,7 @@ namespace Sannel
 	{
 		namespace Sensor
 		{
-			enum class SensorTypes : long
+			enum class SensorTypes : int
 			{
 				Test = 0,
 				Temperature = 1,
